Total price done. 6th frame almost done .The only one that remains is the last frame which is the 7th one...astro chilling a bottle of orange juice fell into me, mid-converstation with the guy that was helping me with the code

+++ src/StudentPackage.java
@@ -13,7 +13,7 @@ public class StudentPackage extends EconomyPackage {
 	public void CalculatePrice(Travelinformation t)
 	{
 		double totalprice;
-		int totalpricewd;
+		double totalpricewd;
 		totalpricewd =Getprice()*(t.getNumber())*(t.getDays());
 		if (Getcar()==true)
 		{
@@ -24,6 +24,7 @@ public class StudentPackage extends EconomyPackage {
 		
 		Totalprice=totalprice;
 		System.out.println(totalprice);
+		
 		
 		
 	}
